u[ps

+++ server/app/Http/Controllers/GeneralConfig/LookUpListController.php
@@ -2,48 +2,44 @@
 
 namespace App\Http\Controllers\GeneralConfig;
 
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use App\Models\GeneralConfig\LookUpList;
 use App\Models\GeneralConfig\LookUpListValue;
-use Illuminate\Http\Request;
 
 class LookUpListController extends Controller {
 
     public function store(Request $request) {
         $query = LookUpList::create($request->all());
         return response([
-            'row'=> $query,
-            'status'=> true
+            'row'=> $query
         ], 200);
     }
 
     public function show(Request $id) {
         return response([
-            'row'=> LookUpList::findOrFail($id),
-            'status'=> true
+            'row'=> LookUpList::find($id)
         ], 200);
     }
 
     public function showAll() {
         return response([
-            'rows'=> LookUpList::get(),
-            'status'=> true
+            'rows'=> LookUpList::get()
         ], 200);
     }
 
     //  Para mostrar los elementos eliminados
     public function getTrashed() {
         return response([
-            'rows'=> LookUpList::onlyTrashed()->get(),
-            'status'=> true
+            'rows'=> LookUpList::onlyTrashed()->get()
         ], 200);
     }
 
     //  Para mostrar un elemento eliminado
     public function recoveryTrashed($id) {
         return response([
-            'row'=> LookUpList::onlyTrashed()->findOrFail($id)->recovery(),
-            'status'=> true
+            'row'=> LookUpList::onlyTrashed()->find($id)->recovery()
         ], 200);
     }
 
@@ -52,12 +48,11 @@ class LookUpListController extends Controller {
     }
 
     public function update(Request $request, $id) {
-        $query = LookUpList::findOrFail($id);
+        $query = LookUpList::find($id);
         $query->fill($request->all())->save();
 
         return response([
-            'row'=> $query,
-            'status'=> true
+            'row'=> $query
         ], 200);
     }
 
@@ -70,9 +65,23 @@ class LookUpListController extends Controller {
     }
 
     public function destroy($id) {
-        $query = LookUpList::findOrFail($id);
-        $query->delete();
+        $exist = DB::table('look_up_lists')->whereExists(function ($query) use ($id) {
+            $query->select(DB::raw(1))
+                ->from('components')
+                ->whereRaw("look_up_list_values.look_up_list_id = $id")
+                ->whereRaw("look_up_list_values.look_up_list_id = look_up_lists.id");
+        })->get();
 
-        return $this->showAll();
+        if(count($exist) < 1){
+            $query = LookUpList::find($id);
+            $query->delete();
+
+            return $this->showAll();
+        } else {
+            return response([
+                'message' => 'Hay un componente vinculado a este grupo',
+                'status'=> false
+            ], 404);
+        }
     }
 }
